requirements.txt
addict==2.2.1
africastalking==1.2.3
asgiref==3.4.1
certifi==2020.6.20
chardet==3.0.4
cloudinary==1.17.0
contextlib2==0.6.0.post1
contextvars==2.4
defusedxml==0.7.0rc1
Django==3.1.4
django-appconf==1.0.4
django-braces==1.14.0 
django-cloudinary-storage==0.3.0
django-cors-headers==3.4.0
django-sistch-drf-stats@git+https://github.com/sisitech/django-common-stats.git@901f5f6
django-distance-field==1.0.3
django-filter==2.3.0
django-imagekit==4.0
django-multiselectfield==0.1.12
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-user-agents==0.4.0
djangorestframework==3.12.2
djangorestframework-bulk==0.2.1
djangorestframework-gis==0.15
django-background-tasks==1.2.5

git+https://github.com/michameiu/drf-autodocs.git@686a788ec0d98986566688a2badf4bc2a698e1ab#egg=drf_autodocs
et-xmlfile==1.0.1
gunicorn==20.0.4
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.7.22
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
immutables==0.14
importlib-metadata==1.7.0
iterable_orm==0.4
jdcal==1.4.1
lxml==4.8.0
Markdown==3.2.2
mock==4.0.3
oauthlib==3.1.0
onesignal-sdk==2.0.0
openpyxl==3.0.4
pilkit==2.0
Pillow==9.1.0
psycopg2-binary==2.9.6
pycparser==2.20
pydot==1.4.1
PyJWT==1.7.1
pyparsing==2.4.7
python-http-client==3.3.1
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rfc3986==1.4.0
schema==0.7.2
sendgrid==6.4.7
six==1.15.0
slack-webhook==1.0.4
sniffio==1.1.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.3.1
starkbank-ecdsa==1.1.0
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
zipp==3.1.0
weasyprint==59.0
pydyf==0.8.0
lxml==4.8.0
python-memcached==1.59
graphviz==0.20.3
python-dotenv==1.0.0
